requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.26,<0.27
django-cors-headers>=3.14,<4.0